Create ANN_hypotetical.R
This code is used to identify the best parameters of the ANN, by generating ANNs on random dates and with random parameters. The dataset is splitted into training and test one, and the random ANN is tested against the latter and the results are then saved to compare with the other simulations' results.